ai-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
aiohttp==3.9.1
websockets==12.0
ta==0.10.2
ccxt==4.1.64
schedule==1.2.0
python-telegram-bot==20.7
asyncio-mqtt==0.16.1
redis==5.0.1
celery==5.3.4
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
yfinance==0.2.28
alpha-vantage==2.3.1

